Configure default gov styles
This introduces an empty homepage controller for demoing
as well as the cells layer for handling views. This provides
a foundation to build the search from.

This commit introduces the cells gem which provides a way to
encapsulate view logic - templates are accompanied with view
models.

Some of the benefits include not relying on rails helpers which
can pollute the view layer as well as testing edge cases more
easily - no complicated controller specs and less heavy feature
specs.