Provide basic feedback on validity of search
This commit introduces the search endpoint and connects
the users input and validates it showing errors when
necessary.

Handling results and no result cases will be introduced
next.

This introduces the concept of using assign_attributes/1
as an alias to Reforms validate/1 method which assigns
values to the properties declared in the form object and
then calls valid on the form. Calling valid on the form
is required to set errors however we can't use a direct
call to valid?/invalid? in the view model as an empty form
which is rendered in the show action would be invalid and
therefore show errors when the user has yet to make a
submission.

This commit also introduces some styles to the search box.