Introduce a detainee details page
This makes use of the DetaineeDetails form object. We
expect that this page/form will be accessible after
clicking the create profile button & for editing from
the profile page(yet to be introduced).

The data will eventually be pulled from a NOMIS api or
copied from a previous escorts detainee record if its
within a certain amount of time(determined by the biz).