Cells Rails 5 testing workaround
A temporary workaround has been put in place to support
Cells with Rails 5. Please check cells/cells-rails for
updates so that the gem can be upgraded and this can be
removed.

Rails 5 no longer supports creating a TestRequest without
providing args to the initializer however we can imitate the
original behaviour using .create