Spec optional fields (#92)
* Add a validates optional field spec helper

* Make rspec aware of the helper

* Example usage

* Remove unnecessary form default

N.B
This work removes the default 'unknown' value on the forms optional/toggle fields. This is something that isn't a Form responsibility as all records have a default value of 'unknown' already set. We also perform a validation that limits the values that can ultimately be persisted to ones explicitly in the list: yes, no & unknown.

* Improve error message on validator

* Apply the validate_optional_field matcher