Remove knowledge of form + schema source of truth
PR Feedback

Pass the dependency to the attribute collection rather than
increase the surface area of the forms api. Nothing really needs
to know about the model and even more so what the model can do.

Benefit of relying on the schemas default values for consistency
when reset is applied.

Please forgive me for mocking the AR object in the spec.