Implement reset functionality (#144)
* Add a collection obj for resetting form attributes

* Implement resetting attributes on the base form

* Spec applying resets only when form is valid

* fixup

* Rubocop mate

* fixup

* Remove knowledge of form + schema source of truth

PR Feedback

Pass the dependency to the attribute collection rather than
increase the surface area of the forms api. Nothing really needs
to know about the model and even more so what the model can do.

Benefit of relying on the schemas default values for consistency
when reset is applied.

Please forgive me for mocking the AR object in the spec.

* Change action method name to reflect procedure

* Use a attr reader for the reset collection

* Assert attribute is reset to default value

For the sake of this test the hash containing the defaults
has been overridden, IRL this should be the models default
column values.