Adds a DB backed session store
We don't render on POST/PUT requests, instead we redirect the user.
In order to get access to invalid form objects on redirects when
invalid we need to access the session. A cookie isn't big enough, we
can have more than 4KB of data in some instances which will raise
a cookie overflow error. DB sessions solve this.