Super-basic analytics framework using exported logs and Kibana (#167)
Introduces LogStash, which replaces the Rails logger with an Elastic Stack compatible JSON implementation. The plan is to use ELK to retrieve basic low-volume behavioural analytics via the included AnalyticsEvent publisher.

The current release of the LogStasher gem is slightly behind master and there's a needed bugfix in the main trunk, so I've had to pin the gem to the commit at master/HEAD as of today. Once the LogStash gem moves beyond release 1.0.1 this can be dropped.


* Testing logstasher

* hmm, improve much?

* LogStasher is yet to release a fixed version, so pinning to master/HEAD for now

* AnalyticsEvent as a proxy for ActiveSupport::Notifications

* Rubocop is the boss of me

* PR feedback