Changes to the Risk from others form
- Update title of first question from 'Rule 45?' to 'Rule 45 (Detainee
separated for their own protection)?'
- Update to when selecting either the 'Yes' or 'No' radio button, do not
display a 'Give details' text field.
- Update title of second question from 'CSRA?' to 'CSRA (Cell Share Risk
Assessment)?'
- Update to when selecting either the 'High' or 'Standard' radio button,
do not display a 'Give details' text field.
- Update title of third question from 'Is the detainee at risk of verbal
abuse from others?' to 'Is the detainee a victim or possible victim of
abuse within prison?'
- On selection of 'Yes' update guide text of displayed text input field
from 'Eg, is the detainee a victim of bullying or high profile in the
media (police or prison officer / celebrity)' to 'Give details'
- Update title of fourth question from 'Is the detainee at risk of
physical abuse from others?' to 'Is the detainee of high public
interest?'
- On selection of 'Yes' update guide text of displayed text input field
from 'Eg, is the detainee a victim of bullying or high profile in the
media (police or prison office / celebrity)' to 'Give details'