Refactor radio_toggle method to support most of its info through options
ATM, in order to define a radio_toggle with specific constrainsts, like
providing a list of specific choices instead of the defaults we were
forced to supply some extra arguments as nil in order to respect the
method arity and arguments order.

The use of options provides a more flexible way of achieving the same
without enforcing us to supply what's optional.