Refactor workflow implementation to use STI
There was a lot of repetition in the way that we were creating/loading a
workflow with a given type.

This refactor removes that repetition and restores the normal use of the
STI approach, so that ActiveModel relationships automatically detect the
type of workflow and load the appropriate class.

NOTE: In order to keep the existent behaviour to have the type as 'risk'
and not 'RiskWorkflow' the workflow model overrides the default
behaviour to determine the type/sti name.