Ensure login errors are being correctly displayed
The login page contained some flash message in the alert severity which
were not being displayed as there was no setup for it until now.

Ths adds the alert severity to the severities that should be displayed
in the page and add a CSS configuration to display it wit the
appropriate style.