Add new questions to the risk of violence section
This adds:
- New columns to the risk table to support the new questions for the
violence section
- Refactor the way we manage the presentation of the violence section so
that it relies on a centralised way (the section object) to determine
how to present the answers for the questions

NOTE: It's becoming increasingly evident the need to avoid the pattern
of adding/removing columns to the database whenever there is new
questions added or questions removed. The idea behind these small
refactors is to gradually migrate to an approach where we have a DSL
that defines the risk/healthcare assessment(s) as a questionaire and the
questions themselves are created dynamically based on that DSL.
Due to the nature of this changes, refactoring this in one go would be
to big of a change, so I'm taking baby steps that leads us to that
outcome.