Centralise mandatory questions for risk assessment
A lot of the information regarding the risk assessment was spread across
models, forms, views, presenters, etc.
This changes centralise all the mandatory questions in each of the
relevant sections for the risk assessment and defines a method in the
risk workflow that retrieves all the mandatory questions for all the
sections.

For now, given that the Risk model is still the one responsible to
determine which questions where answered, I choose to keep the behaviour
only by delegating the mandatory questions to the risk workflow.