Support for both localisation ofradio button labels and options
At the moment, the current form builder we're using does not support
that a radio button label and its options are both localised at the same
time. It also does not allow us to define the label for the radio button
appropriately, if we try to add a radio button fieldset without a label,
the current code detects that as not present and defaults to humanised
version of the attribute.

Unfortunately, some of these changes have been suggested and PRs created
for the govuk_elements_form_builder, but no action has been taken, so
rather than waiting eternally for this changes to be added to the gem,
the MPS custom form builder that inherits from it ends up overriding
some of this methods to provided support for the required functionality.

This is not ideal, as we are ending up repeating code just to be able to
change one or two lines, which also triggered some Rubocop offences that
due to the reasoning provided aboved, I chose to mark the offended file
to be ignored for the time being.

Once these changes take place in the upstream version of the gem I'm
more than happy to get rid of this repetion of code.