Refactor the detainee profile page
So far the profile page was accessed by supplying the move identifier.
The introduction of the breadcrumbs flagged that there's some occurences
where the move doesn't yet exist, but the user has a breadcrumb that
allows him/her to access the profile page. For this reason, and to
introduce some consistency in terms of routes, the profile page is now
based on the detainee identifier and by default it will display
information about the most recent active move (if exists).

If there's no current active move, the profile page will show a link
that allows the user to create a new move.