Move all detainee actions into a detainees controller
The actions performed to a detainee so far were spread among a set of
controllers and the naming for those controller was not very clear.

These changes centralise all the detainee actions in its own controller
and falls back to the conventions use for a given resource:

- The previous profile page can now just be accessed by requesting the
show page as /detainees/:id
- The forms to create and edit a detainee follow the CRUD conventions,
so:
  - GET /detainees/new renders the form to create a new detainee
  - POST /detainees creates a new detainee
  - GET /detainees/:id/edit renders the form to edit an existent
detainee
  - PUT /detainees/:id updates an existent detainee

The changes (IMO) bring a lot more clarity to the actions we support and
remove unnecessary complexity.