Handle checkboxes in IE8 with onchange events
Currently IE8 does not support change events on checkboxes.
The current interface to add destinations to the move hides the
checkboxes, and the change of the value of the checkbox is triggered by
clicking on the associated label.

Due to this fact, whenever we click on the label IE8, does not recognise
the change event, meaning that it never updates the checkbox.

Related to:
http://www.ridgesolutions.ie/index.php/2011/03/02/ie8-chage-jquery-event-not-firing/

These changes ensures that IE8 behaves like the other browsers.