Add date picker to date field in move form
These changes re-use the work done in the homepage where a datepicker and
a radio today/tomorrow were added.

- Extend MPS form builder to support date picker field. The current
implementation of the text_field in the UK Gov Form Builder is too
restrictive and does not allow to customise the way the fields are
displayed, specially if there's extra DOM elements required to be
displayed. For this reason, I have extended the MPS form builder to
provided that support. The downside of it, is that in case of an upgrade
of the UK Gov Form Builder, the MPS forms might no longer work
appropriately (that being said, that already happened ever since the MPS
form builder was created). So, in summary, any upgrades to the UK Gov
Form Builder might require an update on the MPS form builder.
- Add CSS to the date picker to be used accross all forms that follow
the UK Gov default display style.