Update detainee information present in move printout
These changes include:
- Update information displayed in the move printout related with the
detainee according to new requirements.
- Minor refactoring in the print routes. They now fall under the moves
namespace to be very clear what the print is all about.
- Add a controller to server detainee images. THe background for this is
the fact that the images are being stored as a blob in the database and
trying to directly render them in the src attribute of an image DOM tag
has some issues, specially in IE8, where there's a limitation in size
for the image data being rendered. So to solve this problem overall, the
application serves as an endpoint the image associated to the detainee.
- Add a detainee presenter to the move printout to encapsulate all the
printout logic related with the detainee.