Display missing text on summary if unanswered
Prior to this fix it was possible to visit a summary page and see “No”’s
for questions that do not have their top level question answered.

For instance:

[TopLevelQuestion] Violence due to discrimination?(Unchecked)
  [NestedQuestion] Risk to females(Unchecked)
  [NestedQuestion] Homosexuals(Unchecked)

And on the summary we would see:

Violence due to discrimination
Risk to females NO
Homosexuals     NO

When in fact we want to see:

Violence due to discrimination
Risk to females Missing
Homosexuals     Missing

Performing a nil check on the question solves this. This doesn’t appear
to have been an issue before as a lot of the top level questions have
defaults in the schema to “unknown”. However new fields have since been
introduced without a default - this is deliberate*. It is believed that
unanswered questions should not have an answer. It's not great having 2
different types of answer but work is planned to change the modelling
of the questions which will remove all traces of "unknown".

* See comment in:
/app/models/concerns/questionable.rb
git commit: f54e76bd006ec8f