Return PDF document instead of HTML with print css
We will eventually want to store the PDF’s as the source
of truth a given move whilst piloting.

This updates the feature spec to assert on the contents of the
PDF as opposed to the html document we send to the PDF generator.

PDF Reader is able to give us back the contents as text as
it would appear reading from top to bottom. Also tested PDF
inspector with little success as content was jumbled.

Adjusted the photo unavailable text on PDF as it was showing
up outside of the image container. The Yes/No column has also been
padded slightly.

We are aware of issues with the PDF reader gem. The more sections we
combine the more brittle these tests are becoming especially with small
style changes.