Remove overlapping table header on PDF
There is a problem with the table header overlapping text after the
first page of details, this commit removes that for now. We will have
to come up with a way to add the detainee identifier to pages eventually.

This commit updates the specs. We need to come up with a better way of
testing this. The more content we have the harder it is. Right now if
we make a change to the PDF updating the spec becomes a time sink.