Introduce a session controller
End goal here is to remove Devise form the app. User management will be
handled by the SSO. We may still want to have persisted users to hook
up to the workflow models and not carry the various user data in the
session.

Utilise the existing omniauth hash helper on the User model to set the
user reference in the session.

The new session template models the current users/sign_in template in
terms of having the same link to the SSO. This landing page might not
be necessary however it should help get the existing specs to pass easily.