Ensure working app when using SSO mock mode
When in development mode the app should still be usable. More
specifically, when logging out, the app should redirect to the 'Login as
SSO' screen (instead of trying, unsucessfully, to redirect to the SSO
app login screen).

In test mode the redirect to SSO app doesn't work by default as well.
These changes introduce a dumb rack server that accepts the requests
that would normally end up in the SSO app.