Remove duplicate aliases retrieved from NOMS API
Note: The data the NOMS API doesn't seem to be normalised, so there's a
lot of information that we retrieve that is not well formatted or
contains duplicate information. The detainee aliases is one of these
examples where we get a lot of duplicates.

The API should be really taking care of this issues, but until that
happens we need to handle this at the client level.