Use pdftotext to extract text from pdf under test
The algorithm pdf-reader uses to detect lines of text has some
shortcomings. The output can appear with extraneous whitespace.
Pdftotext produces better output so we’ve switched to that.

This method of testing isn’t the greatest, layout changes will alter
the test results. This has been useful in flagging some issues such
as the different fonts used in the different environments(OSX vs linux).
It might be better to test the HTML we pass to webkit html to pdf.