Add headers to PDF printout pages
Note: The tool to generate the PDF is a bit flakey. Any minor changes to
the layout seem to have a major impact in the display, some even making
the PDF to fail to render.

There's some known display issues, maybe directly or indirectly related with
the addition of the headers.