Introduce an offences fetcher
This is the underlying tool to call the charges endpoint in NOMIS and
transform the result from a list of bookings into a list of active
offences.

The API needs some work to make it sensible. We call charges, receive
bookings which is an array containing nested legal cases which in turn
contain the charges which contain the offences.

From a business logic POV we only want to look at bookings with an
active status. Then for each legal case that is active we want to extract
all of the offences from all of the charges. An offence in our domain
is a description and the case reference its associated with. The case
reference in NOMIS is known as the case info number found in a legal
case object.

A lot of this code has followed the existing patterns in the other
fetcher service objects(Detainee & Image). A bit of a hack has been in
place to alias the FetchedResponse’s to_h to data. In the case of this
fetcher we return a list not a hash.