Introduce fetching of offences into front end
Offences should only be called when displaying the show page and only
when no current offences exist for a detainee.

This commit only satisfies the happy path. It appears as though other
parts of the app that interface with Nomis as part of the happy path
Stub the api to return a 404 in the UI this will force manual input.
This commit follows the same pattern.