Refactor data modelling to guarantee idempotency of escort
As it standed everytime a new escort was created the information used
for a previous escort, namely, detainee information, risk assessment,
healthcare assessment, offences would be re-used and most likely
modified which meant we would track of the information used to perform
the previous escort.

These changes ensure that all the information used in an escort are
related only to that escort and cannot be modify outside the scope of
the escort itself.

This turns out to be quite a big change in terms of:
- Routes - all the data required to be created/modified is now under the
escorts scope
- Detainee and Move are not longer directly related but instead they
belong to an Escort record. There's potential, in the future for these
two models to be normalised and all the information related with them to
be present directly in the Escort record. Due to the nature of the
refactor we have refrained from doing any more changes than the MVP.
- The risk and healthcare assessments and the offences are currently
related to a detainee, so the escort record access its information by
delegating their call to the detainee. There's potential for this to be
move directly to the escort, so that the escort has a directly
relationship with the data it requires
- The workflows associated with the move, the risk and healthcare
assessments, and the offences, are defined in a very convoluted way
where the detainee holds the relation to the data but the move holds the
relation to the status of the data. In this PR we masked this by having
the models themselves to response to status enquires. Under the hood, we
are still navigation to this convoluted way of accessing the status. In
a future iteration we will simplify the process by having the model to
respond directly to this status enquires.
- This PR starts to establish the responsability to the status of the
escort to the escort itself, although it currently does it by delegating
some of its status checks to the Move record. This is a transiction step
as the Move record has a lot of responsability that shouldn't really
have, but again, refactoring this would increase even more the size of
the PR which is already quite big.

Future iterations:
- Establish correct relationships between models
- Have escort to control directly its status