Use offences vocab across the application
Prior to this work we had the idea of past & current offences with the
removal of past offences we had both offences and current offences which
are both the same thing.

Maintains references to current offences in the UI - we have some copy
changes dropping real soon - changing this might cause some confusion
when those changes get applied.