Add ability to retrieve application version info
It’s useful across the environments to be able to see what has been
deployed and when it was deployed.

The information returned is consistent with what other apps in the
department return.

With regard to the environment variables we expect here(e.g APP_GIT_COMMIT)
they have been lifted from the ci-deploy repo.