Render risk forms from assessment schema
This is a first step into a new approach of managaging assessments.

This step adds a new schema which will be used in the future to describe
all the information for the available assessments.

For the moment the schema is used only to render the existent assessment
form for the risk of violence section.

In the future it will also be used to:
- determine how the data submitted from the form should be validated
- create/update an answer record for each of the questions answered