Render healthcare forms from assessment schema
This is the second step into a new approach of managing assessments.

This new approach adds a schema which will be used in the future to 
describe all the information for the assessments.

The work in this PR adds the healthcare information into the assessment 
schema and uses it to render the healthcare forms.

The work in this PR contains also:

- Move shared field partials in app/views/shared
- Add fields_for shared partial to manage multiple medications
- Move select_values_for helper in ApplicationHelper