Simplify escort model relations
This is another step towards simplifying the data model that we have at
the moment by removing any unnecessary indirection.

Currently to assert the escort's risk/healthcare assessment
content/status we need to traverse through the detainee model to achieve
this. Although the risk/healthcare assessment are related to the
detainee, they're actually represent an assessment for this particular
escort (and an escort is already related with a detainee).

So, this PR includes the following changes:

- Move risk/healthcare assessment relations from detainee to escort
- Remove automatic creation of risk/healthcare assessment dummy record
when detainee is created
- Create risk/healthcare assessment only when user specifically
initiates the assessment for the escort
- Move risk/healthcare summary pages to risk/healthcare show views