Add overwrite_with_actual_content helper
Every time that we need to record the actual content of the
generated PDF file we were manually adding code to do this matcher.
With this helper, passing an optional parameter to the helper will
regenerate the content.

This follows the same approach as the VCR gem to record new content.