Use assessments schema to render assessment summary page
This is yet another step to centralise the assessment management.

This step will replace the existent method to render the assessments
summary page which was using the assessment section specific classes.

Instead it will use the centralised assessment schema to render that
same information.

With this in place the following is now using the assessment schema:

- render the assessment section form page
- validate the answers submitted for an assessment section
- render the assessment printout
- render the assessment summary page