Extend functionality for displaying assessment answers
The current implementation had a limitation when the question being
processed contained an answer that requires an additional group of
questions to be answered, the display would only render that group of
questions.

The new functionality, mantains the condition (question contains an
answer that requires an additional group of questions), but instead of
displaying only that group of questions, it will display any questions
that the answer depends upon.

As an example:

- Question 1 is answered with Answer A.
- Answer A requires two additional questions to be answered, Question 2 and Question 3.
- Question 2 is just a text field, so doesn't require any additional
information.
- Question 3 is of type group and contains 2 Questions, Question 3G1 and
Question 3G2, both just boolean fields.

The presentation of Question 1 in the summary and print out given the
previous question definition will be:
- Does not display Question 1
- Displays answer for Question 2
- Does not display answer for Question 3
- Displays answers for Question 3G1 and Question 3G2