Remove indirection from escort to move models
Until now asserting the state of the escort was being delegated to the
move due to the phased migration from the legacy data model.

From now on, escort will respond directly to its current status.

This also includes the following changes:

- Whenever an escort is being created by using a clone of an existent
escort for the same prisoner, the move details are no longer copied
across and have to be submitted as it the user was creating a new
escort. The remaining details (detainee details, risk and healthcare
assessment and offences)