Manage status inside risk and healthcare
We want to manage the status changes of assessments inside the assessments
as this will simplify things, so we added a status column to Risk and
Healthcare models.

In this work we also got rid of the not_started state for assessments.

This is the first step in getting rid of the concept of Workflow