Update hint wording for healthcare medical contact
Currently there's little support for dynamic information within the
locales, which prevent us to do something for flexible like in this
example where the wording should reflect the current establishment the
prisoner is at.

The wording got updated with a hard-coded prison (which is the one being
used for the pilot) until we re-visit this requirement again when the
application expands to multiple prisons.