Make offences data read only for issued PERs
So far, after the PER had been issued it was still possible to attempt
to change the offences information (although, once the form was
submitted an error would be shown).

These changes ensure that the user can no longer attempt to change that
information, and that information is displayed in a table format rather
than in a form as they were until now.

Took the opportunity as well to do some small refactoring in the specs
for the assessment summary pages and extract their common functionality
into a mixin.