Refactored assessment complex attributes management
These changes include:

- Manage all the complex attributes that an assessment has just like
another attribute on the main table, with the exception that the
attribute is serialized as an array and may contain zero to many
records. This used to be a relationship that had to be defined to the
main model for each of the complex attributes we had (so far,
medications and must not return establishments).
- Delete medications table due to the above
- Update form rendering so that both Risk and Healthcare assessment
forms are rendered using the assessment form classes.
- Remove specific form classes for Healthcare due to the above
- Update healthcare assessment schema with the validations required for
each of questions in it
- Add must_not_return details (complex attribute with establishment and
establishment details) to risk assessment.

NOTES:
- Admittedly there's still some complexity that I would like to
remove, more specifically in the form classes. Now that we're converging
to a common way of managing assessment based on well define conventions,
that next step should be a lot easier.
- There's some know warnings triggered by some of the classes using
SimpleDelegator (on the method :to_ary), which makes consider that we
might need to take another strategy for delegating behaviour.
- Due to the 2 notes above, I've choosen to mark some of the more
complex files in rubocop to be ignored so they don't trigger any
offences. This will be re-visited ASAP.