Replace the use of Date.today with Time.current.to_date
This is due to the fact that Date.today does not use the application
time zone, but the system local time instead
(https://stackoverflow.com/questions/26952421/whats-wrong-with-railss-date),
where as Date.tomorrow does which can lead to unexpected date
calculation issues.

These changes make use of Time class to produce a date value always
based on the time for timezone the application is set.