Add rake task to generate and store PER documents
This tasks will only generate and store PER documents for issued PERs
that haven't yet stored the generated document.

This is to bring any existent escorts up to date with the new behaviour
of producing PER documents: they're generated once when they're issued,
rather than being generated on the demand like it was happening until
now.