This fixes a minor migration deprecation warning
In the new versions of Rails when a new migration is created the version
of the Rails version that is being used at the time is attached to the
migration. The generator used by the paperclip gem is still not using
that approach which causes the app to show a deprecation warning.