Fix Cross-site scripting issues reported in the pen test
Most of the issues reported have to do with the fact that, normally, the
rails views sanitize the fields by default, but in these cases the
values were set in presenters where none of them were being sanitized.

This might required a more broader look at the use of presenters to
ensure any value that requires sanitize does get sanitize.