Ensure complex attr forms responde to complex attribute methods
Due to the way complex attributes are managed at the moment, we need to
ensure the complex attr form responds to the complex attributes methods.

Going forward, the complex attribute object should probably be loaded
with all the defined complex attributes so that this specific delegation
is no longer required.